chore: backend - eslint updates , split app and server in to independent parts
