+++ packages/server/package.json
@@ -20,6 +20,7 @@
     "eslint-import-resolver-typescript": "^2.5.0",
     "eslint-plugin-import": "2.25.2",
     "eslint-plugin-prettier": "4.0.0",
+    "eslint-plugin-security": "^1.4.0",
     "lint-staged": "11.2.6",
     "nodemon": "2.0.14",
     "prettier": "2.4.1",
